Uber reformatting and massive content/code scope++
This is a massive commit that started off as a commit point with simple
reformatting (syntax mostly, some grammar) and de-list-ing lists that
aren't lists. However, it very quickly expanded to include:

- Vertical block style alignment and list rewriting
- Addition of some flow content
- Delete duplication of rules
- Build tool fixes and expansions

Overall, while extensive in changes, the above list (in theory) does not
change any rules in terms of functional equivalence standards.
However...

This commit also contains a number of rule book changes. All of these
were conducted after consultation with the other members of the project,
and I followed whatever was the consensus.